Give route same type as handlers, add fallthrough.
`route` now returns an `IO (Maybe Response)`, which allows for it to
fail completely. Any defaulting should happen explicitly and outside of
the context of routing, as routes should be able to be nested.

`fallthrough` takes an `IO (Maybe Response)` and a default `IO Response`
and returns an `IO Response`. This allows (and forces!) you to set a
404 handler before calling `toWAI`.