Move example to own .cabal file; expand README.
When it was all in one cabal file, even if you were only using it as a
library it would pull in all the dependencies of the example, which is
obvious not desired.

Also, changed the `notFound*` helpers to return `Response` instead of
`Maybe Response` as they will most likely be used with `fallthrough`.