Add file uploads support; create our own FnRequest
We can only parse the request body once (since it is a conduit source),
so we need to be able to pass the results along, and there is nowhere in
the `Request` to store them. So our `FnRequest` is just a normal WAI
request and space for the parsed post body.

This also introduces `file` and `files` matchers - the first looks for a
file by parameter name, the rest passes all files along.