Fix GHC 7.8.4, make circle build 7.8.4 & 7.10.2.
There were missing imports of `<$>`, which this fixes. Additionally,
some more packages needed to be specified in the stack.yaml because they
are missing from the lts-2.22 snapshot (which is how we're testing 7.8.4
on circle). Since there were only a few changes, we do this and specify
the resolver on the command line rather than having separate stack.yaml files.